requirements.txt
asn1crypto==0.24.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
Django==2.2.5
django-cors-headers==3.1.0
django-extensions==2.2.1
django-phone-field==1.8.0
django-rest-knox==4.1.0
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
httpie==1.0.3
idna==2.8
pycparser==2.19
Pygments==2.4.2
PyJWT==1.7.1
pytz==2019.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
